pyproject.toml
[project]
name = "campsite-crawler"
version = "0.1.0"
description = "Poll recreation.gov for campsite availability and send push notifications"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.scripts]
campsite-crawler = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]
